// UI/src/app/models/policy-model.ts
export class PolicyModel {
    policyId: string;
    dateOfPurchase: string;
    customerId: string;
    fuel: string;
    vehicleSegment: string;
    premium: string;
    bodilyInjuredLiability: boolean;
    personalInjuryProtection: boolean;
    propertyDamageLiability: boolean;
    collision: boolean;
    comprehensive: boolean;
    customerGender: boolean
    customerIncomeGroup: string;
    customerRegion: string;
    customerMaritalStatus: boolean;
}